move class removed a lot of unecesarry bits

+++ chess_game_cpp/src/Move.hpp
@@ -8,19 +8,19 @@ class Position;
 class Move {
 public:
   Move() = default;
-  Move(int from, int to, int flags) : moveNum(from | (to << 6) | flags << 12){};
-  Move(int from, int to, int flags, int score)
+  Move(uint from, uint to, uint flags)
+      : moveNum(from | (to << 6) | flags << 12){};
+  Move(uint from, uint to, uint flags, uint16_t score)
       : moveNum(from | (to << 6) | flags << 12 | (score << 16)){};
 
   // Setters;
-  void setScore(const int &score);
+  void setScore(const uint16_t &score);
   // Getters;
   uint32_t getMoveInt() const;
-  int getTo() const;
-  int getFrom() const;
-  int getFlags() const;
-
-  int getScore() const;
+  uint getTo() const;
+  uint getFrom() const;
+  uint getFlags() const;
+  uint16_t getScore() const;
   // returns noPiece for non captures;
   bool isCapture() const;
